requirements.txt
# 基础依赖
numpy>=1.18.5
opencv-python>=4.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# YOLOv5依赖
matplotlib>=3.2.2
Pillow>=7.1.2
seaborn>=0.11.0
pandas>=1.1.4

# 导出格式支持
# onnx>=1.9.0  # ONNX导出
# onnx-simplifier>=0.3.6  # ONNX优化
# tensorrt>=8.0.0  # TensorRT导出和推理（可选，建议在妙算上使用）

# 开发工具
ipython  # 交互式开发 